Add source_ops.rb file
