+++ app.rb
@@ -12,32 +12,24 @@ require_relative 'book_ops'
 require_relative 'music_album_ops'
 require_relative 'genre_ops'
 require_relative 'movie.ops.rb'
+require_relative 'source_ops.rb'
 
 class App
-  attr_reader :genre, :aops, :lops, :gops, :maops, :gnops, :mops
+  attr_reader :genre, :aops, :lops, :gops, :maops, :gnops, :mops, :sops
 
   def initialize
     # @authors = []
     # @games = []
     # @books = []
     # @labels = []
-    @movies = MovieOps.new
     @aops = AuthorOps.new
     @gops = GameOps.new
     @lops = LabelOps.new
     @bops = BookOps.new
     @maops = MusicAlbumOps.new
     @gnops = GenreOps.new
-  end
-
-  def list_all_sources
-    if @sources.empty?
-      puts 'No sources added!'
-    else
-      @sources.each_with_index do |source, idx|
-        puts "#{idx + 1}) Source: #{source.name}"
-      end
-    end
+    @movies = MovieOps.new
+    @sops = SourceOps.new
   end
 
   def exit
@@ -61,14 +53,14 @@ class App
       4 => -> { @lops.list_all_labels },
       5 => -> { @mops.list_all_movies },
       6 => -> { @mops.create_movie },
-      7 => -> { list_all_sources },
+      7 => -> { @sops.list_all_sources },
       8 => -> { @aops.create_author_with_banner },
       9 => -> { @gops.create_game_with_banner(app) },
       10 => -> { @bops.create_book(app) },
       11 => -> { @lops.create_label },
       12 => -> { @maops.list_album_with_banner },
       13 => -> { @maops.create_album_with_banner(app) },
-      14 => -> { puts 'create a source' },
+      14 => -> { @sops.create_source },
       15 => -> { @gnops.list_genre_with_banner },
       16 => -> { @gnops.create_genre_with_banner },
       0 => method(:exit)
